Reorganize github_async into a package with lazy model imports
